+++ packages/placement-ordering/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.7"></a>
+## [1.0.7](https://github.com/PieELements/pie-elements/compare/@pie-elements/placement-ordering@1.0.6...@pie-elements/placement-ordering@1.0.7) (2018-01-24)
+
+
+### Bug Fixes
+
+* **dependencies:** bump deps ([c61391d](https://github.com/PieELements/pie-elements/commit/c61391d))
+
+
+
+
+<a name="1.0.6"></a>
+## [1.0.6](https://github.com/PieELements/pie-elements/compare/@pie-elements/placement-ordering@1.0.5...@pie-elements/placement-ordering@1.0.6) (2018-01-24)
+
+
+
+
+**Note:** Version bump only for package @pie-elements/placement-ordering
+
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/PieELements/pie-elements/compare/@pie-elements/placement-ordering@1.0.4...@pie-elements/placement-ordering@1.0.5) (2018-01-17)
+
+
+
+
+**Note:** Version bump only for package @pie-elements/placement-ordering
+
+<a name="1.0.4"></a>
+## [1.0.4](https://github.com/PieELements/pie-elements/compare/@pie-elements/placement-ordering@1.0.3...@pie-elements/placement-ordering@1.0.4) (2018-01-11)
+
+
+
+
+**Note:** Version bump only for package @pie-elements/placement-ordering
+
 <a name="1.0.3"></a>
 ## [1.0.3](https://github.com/PieELements/pie-elements/compare/@pie-elements/placement-ordering@1.0.2...@pie-elements/placement-ordering@1.0.3) (2017-12-15)